updated favorite button
